gcc_64bit_ogre_resource.patch applied. Thanks to noisy - fixed GCC compiler warnings - made MyGUI to compile for 64-bit systems - resource file loading through the OGRE resource system

+++ MyGUIEngine/src/MyGUI_PointerManager.cpp
@@ -28,7 +28,7 @@ namespace MyGUI
 
 		Ogre::OverlayManager &overlayManager = Ogre::OverlayManager::getSingleton();
 		mOverlayElement = static_cast<SharedPanelAlphaOverlayElement *>(overlayManager.createOverlayElement(
-			"SharedPanelAlpha", Ogre::StringConverter::toString((int)this) + "_PointerManager" ));
+			"SharedPanelAlpha", Ogre::StringConverter::toString((size_t)this) + "_PointerManager" ));
 		// устанавливаем колличество саб оверлеев
 		mOverlayElement->setCountSharedOverlay(1);
